requirements.txt
streamlit
tensorflow
opencv-python-headless==4.8.0.76
numpy==1.26.4
Pillow